% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_paper.R
\name{download_paper}
\alias{download_paper}
\title{Downloads a Paper}
\usage{
download_paper(paper, path = tempfile(fileext = ".pdf"), open = TRUE)
}
\arguments{
\item{paper}{\emph{character}, DOI or URL of the paper you want to download.}

\item{path}{\emph{character}, optional path of the \code{.pdf} file (full, with a
filename), defaults to  temporary file.}

\item{open}{\emph{logical}, whether to open the result (defaults to \code{TRUE}).}
}
\value{

}
\description{
Unofficial API to \url{sci-hub.se}.
}
